Add-local-and -fix
